+++ apps/rps-game/src/pages/api/authz/user.ts
@@ -1,5 +1,5 @@
 import { type NextApiRequest, type NextApiResponse } from 'next'
-import { generateKeys, sign } from '../../../../utils/crypto'
+import { generateKeys, sign } from '../../../utils/crypto'
 
 const userAuthorization = async (req: NextApiRequest, res: NextApiResponse) => {
   if (req.method !== 'POST') {
