+++ package.json
@@ -41,5 +41,8 @@
   "engines": {
     "node": ">=14.0.0"
   },
-  "packageManager": "npm@8.15.0"
+  "packageManager": "npm@8.15.0",
+  "dependencies": {
+    "@onflow/six-create-account": "^0.0.9"
+  }
 }
